Servant.PureScript

Documentation

generatePSModule :: (GenerateList (Foreign api), HasForeign api) => PSSettings -> ST -> Proxy api -> ST

Given a servant api, generate a PureScript module containing a list of functions for AJAX requests.

generatePSUtilModule :: PSSettings -> (ST, ST)

generatePS

Arguments

:: PSSettings
PureScript rendering settings

-> Req
AJAX request to render

-> ST
Rendered PureScript

Generate a single PureScript function for an AJAX request. To prevent conflicts, generates a unique function name for every available function name and set of captures.

data PSSettings

PureScript rendering settings

Constructors

PSSettings

Fields

_baseURL :: String

Base URL for AJAX requests

_utilModuleName :: ST

module that all generated ajax modules depend on

baseURL :: Lens' PSSettings String

defaultSettings :: PSSettings

Default PureScript settings: specifies an empty base URL